Revert "Password reminder page"
