+++ .travis.yml
@@ -1,0 +1,25 @@
+language: node_js
+
+node_js:
+  - 'node'
+  - 'lts/*'
+
+script:
+  - node ./internals/scripts/generate-templates-for-linting
+  - npm test -- --maxWorkers=4
+  - npm run build
+
+before_install:
+  - export CHROME_BIN=chromium-browser
+  - export DISPLAY=:99.0
+  - sh -e /etc/init.d/xvfb start
+
+notifications:
+  email:
+    on_failure: change
+
+after_success: 'npm run coveralls'
+
+cache:
+  directories:
+    - node_modules
