--- app/components/PublicInput/tests/index.test.js
@@ -1,27 +1,0 @@
-/**
- *
- * Tests for PublicInput
- *
- * @see https://github.com/react-boilerplate/react-boilerplate/tree/master/docs/testing
- *
- */
-
-import React from 'react'
-import { render } from '@testing-library/react'
-
-import PublicInput from '../index'
-
-describe('<PublicInput />', () => {
-  it('Expect to not log errors in console', () => {
-    const spy = jest.spyOn(global.console, 'error')
-    render(<PublicInput />)
-    expect(spy).not.toHaveBeenCalled()
-  })
-
-  it('Should render and match the snapshot', () => {
-    const {
-      container: { firstChild },
-    } = render(<PublicInput />)
-    expect(firstChild).toMatchSnapshot()
-  })
-})
