+++ app/containers/Login/index.js
@@ -9,8 +9,7 @@ import PropTypes from 'prop-types'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import * as Yup from 'yup'
 import { Formik } from 'formik'
-import Button from 'components/Button'
-import PublicInput from 'components/PublicInput'
+import { Icons } from 'themes'
 import messages from './messages'
 import * as Styled from './styled'
 
@@ -27,58 +26,85 @@ const validationSchema = ({ intl }) =>
 const Login = ({ intl }) => {
   const submitLoginForm = () => {}
   return (
-    <Styled.Container>
-      <Styled.Box>
-        <Styled.HeaderContainer>
-          <Styled.Avatar size="150" />
-          <Styled.Header>
-            <FormattedMessage {...messages.header} />
-          </Styled.Header>
-        </Styled.HeaderContainer>
-        <Formik onSubmit={submitLoginForm} initialValues={initialValues} validationSchema={validationSchema({ intl })}>
-          {({ errors, touched, values, handleChange, handleBlur, handleSubmit }) => (
-            <form autoComplete="off">
-              <PublicInput
-                id="emailField"
-                icon={Styled.UserIcon}
-                label={intl.formatMessage(messages.emailLabel)}
-                error={touched.email && errors.email}
-                inputProps={{
-                  type: 'email',
-                  name: 'email',
-                  value: values.email,
-                  onBlur: handleBlur,
-                  onChange: handleChange,
-                }}
-              />
-              <PublicInput
-                id="passwordField"
-                icon={Styled.LockIcon}
-                label={intl.formatMessage(messages.passwordLabel)}
-                error={touched.password && errors.password}
-                inputProps={{
-                  type: 'password',
-                  name: 'password',
-                  value: values.password,
-                  onBlur: handleBlur,
-                  onChange: handleChange,
-                }}
-              />
-              <Button type="submit" onClick={handleSubmit}>
-                <FormattedMessage {...messages.buttonTitle} />
-              </Button>
-              <Styled.Link to="/password-remind">
-                <FormattedMessage {...messages.forgotPasswordMessage} />
-              </Styled.Link>
-              <Styled.Link to="/registration">
-                <FormattedMessage {...messages.noAccount} />
-              </Styled.Link>
-            </form>
-          )}
-        </Formik>
-      </Styled.Box>
-    </Styled.Container>
+    <Styled.Box>
+      <Styled.Avatar src={Icons.loginIcon} alt="loginIcon" />
+      <Styled.Header>
+        <FormattedMessage {...messages.header} />
+      </Styled.Header>
+      <Formik onSubmit={submitLoginForm} initialValues={initialValues} validationSchema={validationSchema({ intl })}>
+        {({ errors, touched, values, handleChange, handleBlur, handleSubmit }) => (
+          <form autoComplete="off">
+            <LoginInput
+              icon={Styled.UserIcon}
+              label={intl.formatMessage(messages.emailLabel)}
+              error={touched.email && errors.email}
+              inputProps={{
+                type: 'email',
+                name: 'email',
+                value: values.email,
+                onBlur: handleBlur,
+                onChange: handleChange,
+              }}
+            />
+            <LoginInput
+              icon={Styled.LockIcon}
+              label={intl.formatMessage(messages.passwordLabel)}
+              error={touched.password && errors.password}
+              inputProps={{
+                type: 'password',
+                name: 'password',
+                value: values.password,
+                onBlur: handleBlur,
+                onChange: handleChange,
+              }}
+            />
+            <Styled.Button type="submit" onClick={handleSubmit}>
+              <FormattedMessage {...messages.buttonTitle} />
+            </Styled.Button>
+            <Styled.Link to="/password-remind">
+              <FormattedMessage {...messages.forgotPasswordMessage} />
+            </Styled.Link>
+            <Styled.Link to="/register">
+              <FormattedMessage {...messages.noAccount} />
+            </Styled.Link>
+          </form>
+        )}
+      </Formik>
+    </Styled.Box>
   )
+}
+
+const LoginInput = ({ icon: Icon, label, error, inputProps }) => (
+  <Styled.InputBox>
+    <Icon size="22" />
+    <Styled.Wrapper>
+      <Styled.Input {...inputProps} />
+      <Styled.Label isEmpty={inputProps.value}>{label}</Styled.Label>
+    </Styled.Wrapper>
+    {error && <Styled.Error color="danger">{error}</Styled.Error>}
+  </Styled.InputBox>
+)
+
+LoginInput.propTypes = {
+  icon: PropTypes.object.isRequired,
+  label: PropTypes.string,
+  error: PropTypes.string,
+  inputProps: PropTypes.shape({
+    type: PropTypes.string,
+    name: PropTypes.string,
+    value: PropTypes.string,
+    onBlur: PropTypes.func,
+    onChange: PropTypes.func,
+  }),
+}
+LoginInput.defaultProps = {
+  label: '',
+  error: null,
+  inputProps: {
+    value: '',
+    onBlur: () => {},
+    onChange: () => {},
+  },
 }
 
 Login.propTypes = {
