--- app/pages/NotFoundPage/Loadable.js
@@ -1,7 +1,0 @@
-/**
- * Asynchronously loads the component for NotFoundPage
- */
-
-import loadable from 'utils/loadable'
-
-export default loadable(() => import('./index'))
