--- app/pages/PasswordRemind/Loadable.js
@@ -1,7 +1,0 @@
-/**
- * Asynchronously loads the component for PasswordRemind
- */
-
-import loadable from 'utils/loadable'
-
-export default loadable(() => import('./index'))
