--- app/pages/PasswordRemind/index.js
@@ -1,5 +1,0 @@
-import React from 'react'
-import PasswordRemindContainer from 'containers/PasswordRemind'
-const PasswordRemind = props => <PasswordRemindContainer {...props} />
-
-export default PasswordRemind
