--- app/pages/Registration/Loadable.js
@@ -1,7 +1,0 @@
-/**
- * Asynchronously loads the component for RegistrationPage
- */
-
-import loadable from 'utils/loadable'
-
-export default loadable(() => import('./index'))
