--- app/pages/Registration/index.js
@@ -1,5 +1,0 @@
-import React from 'react'
-import RegistrationContainer from 'containers/Registration'
-const Registration = props => <RegistrationContainer {...props} />
-
-export default Registration
