+++ app/themes/index.js
@@ -1,6 +1,4 @@
 import Colors from './colors'
 import Icons from './icons'
-import Media from './media'
-import Metrics from './metrics'
 
-export { Colors, Icons, Metrics, Media }
+export { Colors, Icons }
