--- app/themes/metrics.js
@@ -1,19 +1,0 @@
-const metrics = {
-  breakpoints: {
-    mobileSm: '360px',
-    mobileNarrow: '400px',
-    mobile: '480px',
-    tabletSm: '600px',
-    tablet: '768px',
-    desktopXs: '960px',
-    desktopSm: '1024px',
-    desktop: '1170px',
-    laptopSm: '1280px',
-    laptop: '1360px',
-    wide: '1440px',
-    xl: '1680px',
-    xxl: '1740px',
-  },
-}
-
-export default metrics
